refactor(ui): implement EditablePageLayout component
